[Index] Call potentially failing functions at the top of the template
Still doesn't really solve the problem of errors happening in the
middle of a page when we've long sent the HTTP 200… but that might
be the right direction.
Can't do anything about repository corruption there without going
full BufferPool, but eh. We should be able to move at least the
vast majority of potentially failing functionality to the top of
every template?

Part of P0010, funded by DTM.