Go fully template-driven, with a single template per page
Now erroring out as early as we can – if we don't have the template,
even before starting the server.
Sure, we have to duplicate the header and footer inclusions on every
page now, but that seems an acceptable compromise for this otherwise
pretty nice architecture.
For now, at least, maybe we'll mess around with nested templates after
all?

Completes P0010, funded by DTM.