[ASM parser] Don't look for instructions or procedures in non-code segments
Because after all, it doesn't make sense to measure position
independence by looking at the data segments – it's those absolute
references in the code segments that make a binary not position-
independent to begin with.

And it shaves off a few hundred milliseconds for now.

Part of P0012, funded by qp.