[Git] Use CommitObjects() to determine the shortest unique commit length
About twice as fast as the previous approach of doing this per-branch…
as long as the repo is in cache, otherwise it's massively slower? 🥴
Since this iterator will only give us every commit once, we also don't
need to test anymore whether we saw a specific commit earlier.

Part of P0143, funded by [Anonymous].