[Git] Look up commits in the abbreviation map before trying to resolve them
Massive performance boost!!!

ResolveRevision() is really slow… but we actually don't need it most
of the time; `master` and `re-baseline` are the only commits we specify
by their branch name, really. Everything else is specified as either a
full or an abbreviated SHA1 hash, and can quickly be looked up in the
abbreviation map.

Part of P0143, funded by [Anonymous].