[Blog] [2022-07-10] Explain the regular bumper bar collision response
The original version of the post explained all the things that could go
wrong in the special cases that the code pointed out, but left the
actual basics unexplained, together with all the common glitches that
they cause. Thanks to touhou-memories for bringing this to my
attention, this was indeed a glaring omission!