[Videos] [Pipeline] Add enough keyframes for decent seeking performance
And watch the compression ratio go to shit. This almost turned into a
rant about AV1 and how an acceptable number of keyframes more than
doubles the size of an encoded video, but AV1 is in fact the *best*
codec here, with baseline encoded sizes that are at least tolerable.
VP8 and *especially* VP9 are actually much worse in that regard.

Part of P0219, funded by Ember2528 and [Anonymous].