[Blog] Keep all video caption text in a single DOM element
This way, we can employ the same grid cell visibility trick that we use
for the videos themselves, which avoids costly layout reflows when
switching, especially on the big /blog page. Sure, we're duplicating
some text there, but the responsiveness gains more than make up for it.

Part of P0222, funded by [Anonymous].