[Videos] [Decoder] Add a web worker for talking to the ZMBV codec
This ensures the main thread is never busy waiting for the ZMBV stuff to
finish, and also allows multiple videos to decode concurrently, without
all the processing time for every video adding together.

It needs two messages to initialize:
- "url" with the URL of the AVI file
- "module" with the WebAssembly module of the ZMBV codec

Once it's initialized, it sends a "ready" message. A frame can be
requested using a "request" message, passing the frame index, the
ImageData's ArrayBuffer and its width and height. The frame is then
copied to the specified buffer. Seeking is handled internally.