[HTML] Add width and height attributes to emoji <img>s
Prevents emojis from blowing up to their intrinsic size. The attribute
is overridden on the website with CSS as intended.

(Added a comment to explain where the value comes from, and the `width`
attribute as well. My favorite test case for limited HTML, `M-x eww` in
Emacs, requires both `width` and `height` for either to have an effect.
--Nmlgc)