[Stripe] Implement subscription cancellation logic
Combining not storing any Stripe IDs on my server with an added layer
of verification. Designed this way because the "thank you" page can
only show the Stripe Checkout session within the next 24 hours after
the completed first payment, which is when the session expires by
default… except that in Stripe terms, "expire" actually means "session
gets frozen in whatever state it's in, and turned immutable"? 🤔
Stripe actually keeps these sessions forever, not least because
they're necessary for the level of detail that Stripe wants to provide
in its transaction logs. Therefore, it makes sense that the API can
retrieve any older session at any point in the future as well:

	https://stackoverflow.com/questions/69600941

Which kind of defeats the point of showing the "password" for
subscription cancellation on a consistently available and technically
public page?
Oh well, it's not like it complicates the UI all too much. It still
makes sense to design with a general subscription management page in
mind, which will prepare us for adding all kinds of Stripe-exclusive
subscription management features at a later point in time.

Part of P0239, funded by Ember2528.