[Videos] [Pipeline] Include audio track presence in metadata
Since only Safari (?!) currently implements `audioTracks`, we'd have to
rely on `webkitAudioDecodedByteCount` on Chrome, which, as the name
suggests, needs the browser to decode part of the video first. Once
again, letting the backend pass this flag to the player is much more
reliable.

Part of P0257, funded by [Anonymous].