[Videos] [Player] Remove AV1 sources on Microsoft Edge
It's completely OK for browsers to not support a video codec. But
while regular and well-behaved browsers just try another one of a
video's sources in that case, Edge insists on staying on a non-playing
AV1 video.
The `codecs` parameter seems as if it would solve this issue:

	https://developer.mozilla.org/en-US/docs/Web/Media/Formats/codecs_parameter
	https://jakearchibald.com/2022/html-codecs-parameter-for-av1/

But it still doesn't change a thing, even when also specifying it for
the VP9 and VP8 sources. Calling the infamously ridiculous
HTMLMediaElement.canPlayType() method with a representative parameter
of "video/webm; codecs=av01.1.04M.08.0.000.01.13.00.0" might explain
why: Both the AV1-supporting Chrome and Edge return `"probably"`, but
only the former can actually play this format. 🤦

What you're supposed to do is to install an extension from the
Microsoft Store:

	https://apps.microsoft.com/detail/av1-video-extension/

But even that stopped working as of Edge version 116:

	https://caniuse.com/av1
	https://www.reddit.com/r/AV1/comments/16mmm35/av1_and_edge/

And even if it still did: If you can't query the presence of this
extension via JavaScript, it might as well not exist at all.

Having a pipeline that encodes every video into two other formats means
that we can at least do *something* about all this. Try visiting
https://github.com/nmlgc/ReC98/releases/tag/P0234-1 in Edge and you
also don't get to see anything, because that video is encoded using
AV1, and GitHub (understandably) doesn't re-encode every uploaded video
into old formats.
Just for comparison, I tried it all on an old Android 6 phone from
2014, and even that phone picked and played the AV1 videos with the
latest available Chrome and Firefox versions. That was the phone that
didn't support VP9 in 2019, which was the initial reason for me to add
the VP8 versions. Looks like it's time to drop those… 🤔
Maybe in the far future once I start running out of space on nmlgc.net.

We could detect Edge via the User-Agent header, but that would just
move us further away from static site generation. Instead, let's nuke
the AV1 sources on the client, immediately before loading each video.

Thanks to touhou-memories for reporting the issue!