[🚚] Deliver P0262-P0263 (TH04/TH05 main/setup menus / TH05 title animation)
Summary: https://rec98.nmlgc.net/blog/2023-11-30

Picking some of the lowest-hanging fruit in exchange for 1.30% overall
finalization within two pushes. We don't have a lot of ASM code like
that left in this codebase…
Also, inverting the colors on that Edge popup so that you can actually
read the link. Not really a fan of that look, but getting the point
across is more important here.

Funded by [Anonymous] and Blue Bolt.