[📸] Shuusou Gyoku: Use WebP screenshots that fit Flathub's listing guidelines
Turns out that Flathub is not kind to this game:

• It displays screenshots at a fixed height of 468 pixels, which has
  got to be some sick joke.
• It converts our lossless PNGs to lossy WebPs, which both adds ugly
  and very noticeable compression artifacts *and* even slightly
  increases their file size 🤣

Since we're only showing off the main menu, we could just crop away
12 pixels at the top. But then again, their app listing guidelines want
that rounded window border and shadow. So let's go for that, ship a
lossless WebP, hope that Flathub doesn't reconvert it, and that it
looks halfway decent on the page.