[Blog] [2025-04-09] Add a second example to the birthday paradox explanation
If we only work out the probability for the maximum amount of colors,
it doesn't immediately follow why this is a problem for images with
slightly fewer colors – which is the actual issue we ran into and which
the slide should have explained. I did provide the exact collision
probabilities in the `<figcaption>`s below, but I don't think it was
clear enough how you would arrive at them.