[Blog] [2023-11-01] Move the face tile caption back into the document flow
<figure>s don't actually need an aspect-ratio. It's only needed for
defining the height of a multilayer <div> that only contains absolutely
positioned children. Without it, that <div> has a height of 0, causing
the <figcaption> to immediately follow and not actually be visible as
the images inside the <div> are rendered on top of it.
This accidentally broke during the multilayer streamlining I did in
20500a5d74045bfddc239d05dab623704999b72f.