[HTML] Replace context-independent ReC98 Twitter links with ActivityPub
Turns out that we only have a grand total of two such Twitter links:
The main one in the header, and the link to the presentation of my
2019 build system, which predates the blog. Every other Twitter link
either points to another user or appears in the context of how much
engagement the tweet received, which we'd lose if we linked to another
platform.

Part of P0329, funded by [Anonymous] and Ember2528.